+++ lib/rubocop/cop/lint/empty_interpolation.rb
@@ -1,0 +1,22 @@
+# encoding: utf-8
+
+module Rubocop
+  module Cop
+    module Lint
+      # This cop checks for empty interpolation.
+      #
+      # @example
+      #
+      #   "result is #{}"
+      class EmptyInterpolation < Cop
+        MSG = 'Empty interpolation detected.'
+
+        def on_dstr(node)
+          node.children.select { |n| n.type == :begin }.each do |begin_node|
+            add_offense(begin_node, :expression) if begin_node.children.empty?
+          end
+        end
+      end
+    end
+  end
+end